added kompanies_warehouses model

+++ migrations/20220224115257_UsersCreationMigration.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "stocs" (
 
-"id" integer not null primary key,
+"id" BIGSERIAL not null primary key,
 "company_sender_id" integer not null,
 "company_recipient_id" integer not null,
 "product_id" integer not null,
