+++ migrations/20220224115331_UsersCreationMigration.up.sql
@@ -1,9 +1,8 @@
 CREATE TABLE "warehouses" (
 
-"id" integer not null primary key,
+"id" BIGSERIAL not null primary key,
 "name" varchar not null,
 "slug" varchar not null unique,
 "company_id" integer not null,
 "address" varchar not null
-
 );
