+++ migrations/20220228120639_UsersCreateKompany.up.sql
@@ -1,9 +1,8 @@
 CREATE TABLE "kompany" (
 
-"id" integer not null primary key,
+"id" BIGSERIAL not null primary key,
 "name" varchar not null,
 "slug" varchar not null unique,
 "inn" integer not null,
 "kpp" integer not null
-
 );
